requirements.txt
firebase-admin
requests
pillow
blurhash-python
google-generativeai